requirements.txt
classnotation==1.5
coverage==5.5
csscompressor==0.9.5
flake8==3.9.2
htmlmin==0.1.12
Jinja2==3.1.4
mypy==1.10.1
pillow==10.3.0
PyAutoGUI==0.9.54
PyYAML==5.4.1
snakeviz==2.1.1
types-PyYAML==5.4.3
watchdog==4.0.1
## The following requirements were added by pip freeze:
MarkupSafe==2.1.5
mccabe==0.6.1
MouseInfo==0.1.3
mypy-extensions==1.0.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==0.1.30
python3-xlib==0.15
pytweening==1.2.0
tomli==2.0.1
tornado==6.4.1
typing_extensions==4.12.2